audio download and transcript/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.3.0
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
language_tool_python==2.9.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai==1.95.1
openai-whisper==20250625
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
regex==2024.11.6
requests==2.32.4
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.9.0
toml==0.10.2
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
yt-dlp==2025.6.30
